+++ AoDDownloader/quality.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+
+class Quality(Enum):
+    bluray = -6
+    fullhd = -5
+    hd = -4
+    dvd = -3
+    sd = -2
+    mobile = -1