+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+# 0.6.0
+- add quality selection
+- save quality to config
+- use the best quality available if max quality is not met
+- get requirements from requirements.txt
+- ensure Python version 3.8 or higher
+
 ## 0.5.0
 - use click for progressbar
 - save language preferences in config